Move run code actions back to liveshare layer so that we can switch to the UI thread.
